refactor(core/biblio): split into pre-processing and rendering parts
